pyproject.toml
[project]
name = "social-media-scraper"
version = "0.1.0"
description = "Unified scraper for Xiaohongshu, Instagram, and LinkedIn"
authors = []
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # Common
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "click>=8.0",
    "loguru>=0.7",
    "aiofiles>=23.0",
    "aiohttp>=3.8",
    "pandas>=2.0",
    # Xiaohongshu
    "PyExecJS>=1.5",
    "retry>=0.9",
    "opencv-python>=4.8",
    "numpy>=1.24",
    "qrcode>=7.4",
    "openpyxl>=3.1",
    "playwright>=1.40",
    # Instagram
    "beautifulsoup4>=4.12",
    # LinkedIn
    "requests>=2.31",
    "lxml>=4.9",
]

[project.scripts]
social-media-scraper = "social_media_scraper.cli:main"

[tool.hatch.build]
sources = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
